Update VotingService.java
Adding and correcting the nullification of a vote. If a person chooses to vote up/down for a comment or post, and their current vote matches the one they try to give, it will cancel out their vote, and also remove their vote of up/down. This will nullify their initial vote.